requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aom=3.6.0=hd77b12b_0
asttokens=2.4.1=pyhd8ed1ab_0
blas=1.0=mkl
blosc=1.21.3=h6c2663c_0
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py310hd77b12b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
certifi=2023.11.17=pyhd8ed1ab_0
cffi=1.16.0=py310h2bbff1b_0
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py310haa95532_0
cloudpickle=2.2.1=py310haa95532_0
colorama=0.4.6=py310haa95532_0
comm=0.2.1=pyhd8ed1ab_0
contourpy=1.2.0=py310h59b6b97_0
cryptography=41.0.7=py310h89fc84f_0
cuda-cccl=12.3.101=0
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.3.101=0
cuda-opencl-dev=12.3.101=0
cuda-profiler-api=12.3.101=0
cuda-runtime=12.1.0=0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.2=py310h2bbff1b_0
dask-core=2023.11.0=py310haa95532_0
dav1d=1.2.1=h2bbff1b_0
debugpy=1.8.0=py310h00ffb61_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=py310haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2023.10.0=py310haa95532_0
giflib=5.2.1=h8cc25b3_3
gmpy2=2.1.2=py310h7f96b67_0
icc_rt=2022.1.0=h6049295_2
icu=73.1=h6c2663c_0
idna=3.4=py310haa95532_0
imagecodecs=2023.1.23=py310h6c6a46e_0
imageio=2.31.4=py310haa95532_0
importlib-metadata=7.0.1=py310haa95532_0
importlib_metadata=7.0.1=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.29.0=pyha63f2e9_0
ipython=8.20.0=pyh7428d3b_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.7.1=py310h5588dad_0
kiwisolver=1.4.4=py310hd77b12b_0
krb5=1.20.1=h5b6d351_0
lazy_loader=0.3=py310haa95532_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libavif=0.11.1=h2bbff1b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.3.4.107=0
libcurand-dev=10.3.4.107=0
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libsodium=1.0.18=h8d14728_1
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py310haa95532_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib=3.8.0=py310haa95532_0
matplotlib-base=3.8.0=py310h4ed8f06_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
munkres=1.1.4=py_0
nest-asyncio=1.5.9=pyhd8ed1ab_0
networkx=3.1=py310haa95532_0
numexpr=2.8.7=py310h2cd9be0_0
numpy=1.26.3=py310h055cbcc_0
numpy-base=1.26.3=py310h65a83cf_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.0=hcfcfb64_1
packaging=23.1=py310haa95532_0
pandas=2.1.4=py310h4ed8f06_0
parso=0.8.3=pyhd8ed1ab_0
partd=1.4.1=py310haa95532_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py310h045eedc_0
pip=23.3.1=py310haa95532_0
platformdirs=4.1.0=pyhd8ed1ab_0
ply=3.11=py310haa95532_0
prompt-toolkit=3.0.42=pyha770c72_0
psutil=5.9.7=py310h8d17308_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.10=py310hd77b12b_0
pyqt5-sip=12.13.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.13=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytorch=2.1.2=py3.10_cuda12.1_cudnn8_0
pytorch-cuda=12.1=hde6ce7c_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py310haa95532_0
pywavelets=1.5.0=py310h9128911_0
pywin32=306=py310h00ffb61_2
pyyaml=6.0.1=py310h2bbff1b_0
pyzmq=25.1.2=py310h2849c00_0
qt-main=5.15.2=h19c9488_10
requests=2.31.0=py310haa95532_0
scikit-image=0.20.0=py310h3513d60_0
scikit-learn=1.3.0=py310h4ed8f06_1
scipy=1.11.4=py310h309d312_0
setuptools=68.2.2=py310haa95532_0
sip=6.7.12=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6c2663c_1
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.4.12=py310haa95532_0
tk=8.6.12=h2bbff1b_0
tomli=2.0.1=py310haa95532_0
toolz=0.12.0=py310haa95532_0
torchaudio=2.1.2=pypi_0
torchinfo=1.8.0=pyhd8ed1ab_0
torchvision=0.16.2=pypi_0
tornado=6.3.3=py310h2bbff1b_0
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=py310haa95532_0
tzdata=2023d=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.18=py310haa95532_0
vc=14.2=h21ff451_1
vc14_runtime=14.38.33130=h82b7239_18
vs2015_runtime=14.38.33130=hcb4865c_18
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.41.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zeromq=4.3.5=h63175ca_0
zfp=1.0.0=hd77b12b_0
zipp=3.17.0=py310haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
